Add author info, and relicense MIT OR Apache-2.0
See #1 for details. For both authors' permission to relicense, see:

- https://github.com/GitoxideLabs/learning-rust-with-gitoxide/issues/1#issuecomment-2795923004
- https://github.com/GitoxideLabs/learning-rust-with-gitoxide/issues/1#issuecomment-2814701661

This states explicitly in the readme that the content authors are
Sebastian Thiel and Sidney Douw. That avoids ambiguity or confusion
that might arise from the commit history (the commits introducing
the content all show Sebastian Thiel only, but they reflect the
collaboration in the LRG videos), or from the repo being moved to
the GitoxideLabs org (which might otherwise lead to the "about"
text "Our sessions when learning Rust with gitoxide" being
misinterpreted).

The readme contained an incomplete list of video topics and commits
that introduced material corresponding to those LRG videos. This
retains that list, but notes that it is incomplete. This is for
general clarity, but also more specifically to avoid giving the
false impression that the newly added authorship information and
license apply to only some of the topics' code.